feat: 번역 함수 개선
