+++ .env.example
@@ -5,3 +5,6 @@ TRANSLATION_MODEL=gpt-4.1
 
 # Deepl 설정
 DEEPL_API_KEY=
+
+# 번역 프롬프트 파일 경로
+TRANSLATION_PROMPT_FILE=translation_prompt.txt
