+++ init.py
@@ -1,137 +1,15 @@
 #!/usr/bin/env python3
 import os
 import shutil
-import subprocess
-import time
-from functools import wraps
 
-import dotenv
-import openai
-from openai.types.chat import ChatCompletionSystemMessageParam, ChatCompletionUserMessageParam
+from dotenv import load_dotenv
 
-
-def retry(max_attempts=3, delay=1, backoff=2, exceptions=(Exception,)):
-    """예외 발생 시 함수를 재시도하는 데코레이터
-
-    Args:
-        max_attempts: 최대 시도 횟수
-        delay: 초기 대기 시간(초)
-        backoff: 대기 시간 증가 요소(다음 시도에서는 delay * backoff)
-        exceptions: 재시도할 예외 클래스 튜플
-    """
-
-    def decorator(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            attempt = 0
-            current_delay = delay
-
-            while attempt < max_attempts:
-                try:
-                    return func(*args, **kwargs)
-                except exceptions as e:
-                    attempt += 1
-                    if attempt >= max_attempts:
-                        print(f"최대 시도 횟수({max_attempts})를 초과했습니다. 마지막 오류: {e}")
-                        raise
-
-                    print(f"오류 발생: {e}. {current_delay}초 후 재시도 ({attempt}/{max_attempts})...")
-                    time.sleep(current_delay)
-                    current_delay *= backoff  # 대기 시간 증가
-
-            return None  # 이 코드는 실행되지 않지만 형식적으로 필요
-
-        return wrapper
-
-    return decorator
-
-
-def timeout(seconds=420):
-    """함수 실행 시 타임아웃을 적용하는 데코레이터
-
-    Args:
-        seconds: 타임아웃 시간(초)
-    """
-
-    def decorator(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            import signal
-
-            def handler(signum, frame):
-                raise TimeoutError(f"함수 실행이 {seconds}초를 초과했습니다.")
-
-            # 타임아웃 설정
-            signal.signal(signal.SIGALRM, handler)
-            signal.alarm(seconds)
-
-            try:
-                result = func(*args, **kwargs)
-                signal.alarm(0)  # 타이머 재설정
-                return result
-            except TimeoutError as e:
-                print(f"타임아웃: {e}")
-                raise
-            finally:
-                signal.alarm(0)  # 타이머 재설정
-
-        return wrapper
-
-    return decorator
-
-
-def run_command(command, cwd=None):
-    result = subprocess.run(
-        command,
-        shell=True,
-        check=True,
-        text=True,
-        capture_output=True,
-        cwd=cwd
-    )
-    return result.stdout.strip()
-
-
-@retry(max_attempts=3, delay=5, backoff=2, exceptions=(Exception,))
-@timeout(seconds=300)  # 5분 타임아웃
-def translate_with_openai(content):
-    """OpenAI API를 사용하여 마크다운 콘텐츠를 영어에서 한국어로 번역합니다."""
-    # OpenAI API 키 가져오기
-    api_key = os.environ.get("OPENAI_API_KEY")
-    if not api_key:
-        print("OPENAI_API_KEY가 설정되지 않았습니다.")
-        return content
-
-    # OpenAI 클라이언트 설정
-    openai.api_key = api_key
-    client = openai.OpenAI()
-
-    # 시스템 프롬프트 설정
-    system_prompt = """당신은 전문 번역가입니다. EN에서 ko로 마크다운 문서를 번역해주세요.
-중요한 지침:
-1. 코드 블록, HTML 태그, 링크 URL은 번역하지 마세요.
-2. 마크다운 형식을 유지하세요.
-3. 전문 용어는 적절하게 번역하세요.
-"""
-
-    try:
-        # 올바른 타입의 메시지 생성
-        system_message = ChatCompletionSystemMessageParam(role="system", content=system_prompt)
-        user_message = ChatCompletionUserMessageParam(role="user", content=f"다음 마크다운 문서를 번역해주세요:\n\n{content}")
-
-        response = client.chat.completions.create(
-            model=os.environ.get("TRANSLATION_MODEL", "gpt-4.1"),
-            messages=[system_message, user_message]
-        )
-        return response.choices[0].message.content
-    except Exception as e:
-        print(f"번역 중 오류 발생: {e}")
-        return content
+from utils import run_command, translate_file
 
 
 def main():
     # .env 파일 로드
-    dotenv.load_dotenv()
+    load_dotenv()
 
     repo_url = "https://github.com/laravel/docs.git"
     branches = ["master", "12.x", "11.x", "10.x", "9.x", "8.x"]
@@ -165,18 +43,11 @@ def main():
             shutil.copy2(source_path, en_target_path)
             print(f"파일 복사: {file} -> {en_target_dir}")
 
-            # 파일 내용 읽기
-            with open(source_path, 'r', encoding='utf-8') as f:
-                content = f.read()
-
             # OpenAI로 번역
             print(f"번역 중: {file}")
-            translated_content = translate_with_openai(content)
-
-            # ko 디렉토리에 번역된 파일 저장
             ko_target_path = os.path.join(ko_target_dir, file)
-            with open(ko_target_path, 'w', encoding='utf-8') as f:
-                f.write(translated_content)
+            # 번역 실행
+            translate_file(source_path, ko_target_path)
             print(f"번역 완료: {file} -> {ko_target_dir}")
 
         print(f"{branch} 처리 완료")
