+++ translation_prompt.txt
@@ -1,0 +1,148 @@
+# 문장 번역 시 생각할 것
+
+1.  **읽는 사람이 이해하기 쉽게 작성해주세요.**
+    * 직역하듯이 번역하는 것보다, 약간의 의역이 있어도 괜찮습니다.
+    * 영어 원문이 3문장으로 되어 있다고 해서 꼭 3개의 한글 문장으로 번역하지 않아도 됩니다.
+    * 좀 더 풀어서 쓰거나, 문장 순서를 바꿔서 번역해도 됩니다.
+    * 많은 경우 문장 앞뒤를 바꾸면 좀 더 읽기 편한 한국어 문장이 됩니다.
+    * 설명된 내용이 어렵다면, (역자주)를 추가해서 배경에 대한 설명을 더 해주셔도 됩니다.
+2.  **용어의 통일**
+    * 영어 번역의 어려움에는 용어의 문제가 있습니다.
+    * 용어집을 꼭 한번 읽어보신 뒤에 번역 작업을 진행해주세요.
+3.  **목차와 링크는 번역하지 않습니다.**
+    * 문서의 초반에 나열되는 목차와 링크는 번역하지 않습니다.
+    * 이 이유는 웹사이트에서 한글 문서를 확인하다가, 도메인을 laravel.kr → laravel.com 으로만 바꾸면 그대로 영어 원문을 확인할 수 있게 하기 위함입니다.
+4.  **번역하기 어려운 경우 영어-한글, 한글-영어 병기 해도 좋습니다.**
+    * 한글번역 하다가 용어가 번역하기 어려운 경우 영어-한글, 한글-영어 형태로 병기해도 좋습니다.
+    * 예를 들면 request-요청, queue-큐 가 이런 경우에 해당합니다.
+5.  **복수형 표현을 한글로 표기할 때는 단수로 표기하는 것을 고려해보세요.**
+    * 예를 들어 “options”를 번역할 때 "옵션들"이라고 하면 어색할 수 있습니다. "옵션"이라고 단수 표기해도 이해하기 쉬울 수 있습니다. 단수 표기를 고려해보세요. 한글은 영어와 달리 "복수", "단수" 표현을 딱 달라 구분하지 않아도 이해하는데 무리가 없는 경우가 많습니다.
+6.  **코드 안에 들어 있는 주석은 꼭 번역하지 않아도 됩니다.**
+    * 소스 코드 안에 들어 있는 주석은 번역이 힘들면 하지 않아도 괜찮습니다.
+    * 번역 하셔도 되고, 안하셔도 무방합니다.
+7.  **종류를 나열하는 문장의 경우 or, and 를 그냥 생략하세요.**
+    * ‘A’, ‘B’, ‘C’, and ‘D’ 와 같은 영어 문장을 번역하면서 ‘가', '나', '다', 그리고 ‘라’ 와 같이 번역하지 말고 그냥 ‘가', '나', '다’, ‘라’ 와 같이 번역하세요. 굳이 "or", "and”를 번역하지 않아도 문장을 이해하는데 문제가 없습니다.
+8.  **Note , Warning 은 다음과 같이 번역합니다.**
+    * > Note 는 > **참고**
+    * > Warning 은 > **주의**
+9.  **고유명사인 경우 그냥 영어 표기합니다.**
+    * Artisan, Scout, Vite, 과 같이 라라벨에서 하위 패키지에 대한 문서의 경우 고유명사를 번역하기 어렵습니다. 그냥 영어 단어 그대로 표기하세요.
+10. **문장이 콜론 (“:”) 으로 끝나는 경우는 한글로 번역할 때 마침표 (“.”) 으로 바꿔주세요.**
+    * 영어랑 달리 한글에서는 마침표로 종료되는 문장이 더 자연스럽습니다.
+11. **md 문서 안에 자체적인 스타일링을 위해서 존재하는 `<style>` 태그는 모두 제거합니다.**
+    * 웹사이트에서 렌더링 될 때 사용하기 위한 `<style>` 태그는 삭제합니다.
+    * `<style>` 태그 안에 있는 css 스타일링 클래스들도 모두 제거합니다.
+12. **마지막으로 한글 번역은 많은 사람들의 기여로 완성됩니다.**
+    * 미흡한 점이 발견되더라도 너그럽게 이해해주시고 '가이드'에 맞게 직접 수정해주세요.
+    * 누구나 실수 할 수 있습니다.
+
+# 용어집
+
+* **laravel**: 한글 표기는 "라라벨" 로 통일합니다. 코드 상에 표시되지 않은 이상 문장에서 언급된 경우는 모두 "라라벨" 이라는 한글 표기로 변경해주세요. 고유명사 취급합니다.
+* **asset**: “애셋” 으로 번역합니다. "자산"으로 번역하지 않습니다.
+* **assert**: "assert" 그대로 번역하거나, '확인하다', '검증하다' 등으로 의역해도 좋습니다. 다른 매뉴얼에서는 '단언하다' 로 표현하기도 합니다만. 용어가 어려워 잘 사용하지 않습니다.
+    * "assertion" 의 경우도 그대로 "assertion"으로 표현합니다.
+* **application**: "애플리케이션" 으로 번역합니다. "응용 프로그램" 으로 번역하지 않습니다.
+* **authenticate**: "인증" 으로 번역합니다.
+* **authorization**: "인가" 로 번역합니다.
+* **cache**: "캐시" 로 번역합니다.
+* **column**: "컬럼" 으로 번역합니다.
+* **command**: "명령어"로 번역합니다.
+* **cookie**: "쿠키"로 번역합니다.
+* **configuration**: "설정", "설정하기" 로 번역합니다.
+    * "구성" 으로 번역되기도 합니다. ex) Configuring Nginx Sites → Nginx 사이트 구성
+    * 목적어로 사용될 때 "설정 내역" 으로 번역되기도 합니다.
+* **console**: "콘솔" 으로 번역합니다.
+* **collections**: "컬렉션" 으로 번역합니다.
+* **closure**: "클로저" 로 번역합니다.
+* **customize**: 사용자 정의라고 번역해도 되지만, 문장이 어색해진다면 커스터마이즈, 커스터마이징과 같이 표현해도 됩니다.
+* **dependency injection**: "의존성 주입"으로 번역합니다.
+* **enum**: "열거형" 으로 번역합니다.
+* **error**: "에러”로 번역합니다.
+* **explicit**: "명시적" 으로 번역합니다.
+* **extension**: "확장기능" 주로 php 자체의 기능을 의미할 때 많이 등장합니다. "확장 기능" 으로 번역을 권장합니다.
+* **facade**: "파사드" 로 번역합니다.
+* **fallback**: "대체 방법" 등으로 번역합니다. 상황에 맞게 의역해도 좋습니다.
+* **helper**: "헬퍼" 로 번역합니다.
+    * 문맥상 필요하다면 "헬퍼 함수" 로 번역할 수 있습니다.
+* **introduce**: 직역하면 "소개" 이지만, 좀 더 자연스럽게 표현하기 위해서 "시작하기"로 번역을 권장합니다.
+* **implicit**: "암묵적" 으로 번역합니다.
+* **log**: "로그"로 번역합니다.
+    * “logging” → "로깅" 으로 번역합니다.
+* **method**: "메서드" 라고 번역합니다. (이게 표준 한국어 표기입니다.) "메소드" 라고 번역하지 않습니다.
+* **mutator**: "mutator" 그대로 영문 표기합니다.
+* **migration**: "마이그레이션"으로 번역합니다.
+* **overriding**: "재정의", "오버라이딩" 혼용하여 사용할 수 있습니다. 문장 의미에 맞게 더 적합한 단어를 선택하세요.
+* **parameter**: "파라미터"로 번역합니다.
+* **pagination**: "페이지네이션" 으로 번역합니다.
+    * paginator 는 목적어 대상으로 그대로 영문표기합니다.
+* **queue**: "큐-queue"로 번역합니다.
+* **retrieve**: "조회", "조회하다"와 같이 표현하세요. (검색하다 라고 번역하지 않습니다.)
+    * 경우에 따라 "가져오기" 라고 번역할 수 있습니다.
+    * 주로 DB 관련 문서에서는 "조회/조회하기", HTTP 요청 관련 문서에서는 "가져오기" 로 번역할 수 있습니다.
+* **relationship**: "연관 관계" 로 번역합니다. "관계" 로 번역하지 않습니다.
+* **request**: request-요청 으로 병기하거나, "요청” 으로 번역해도 됩니다.
+    * 단독으로 사용되면 "request-요청" 으로 번역하는 경우가 많고
+    * "HTTP request" 와 같은 표현에서는 "HTTP 요청" 으로 번역하는 경우가 많습니다.
+* **redirect**: "리다이렉트" 라고 번역합니다.
+    * "redirection" → "리다이렉션" 으로 번역합니다.
+* **route**: "라우트" 로 번역합니다.
+    * routing → "라우팅" 으로 번역합니다.
+    * url 경로를 의미하는 경우 "경로" 라고 번역할 수 있습니다.
+* **resolve**: 주로 의존성과 관련되어 많이 사용되는 단어로 "의존성 해결"로 번역할 수 있습니다.
+* **redis**: 고유명사지만 워낙 유명한 라이브러리라 "레디스" 라고 번역하거나 Redis 영문 그대로 사용해도 됩니다.
+* **service container**: "서비스 컨테이너" 로 번역합니다.
+* **session**: "세션" 으로 번역합니다.
+* **shortcut**: "지름길", "단축어" 등으로 표현할 수 있는 단어입니다. 다만 맥락에 따라 "좀 더 편리한 방법" 등으로 의역을 권장합니다.
+* **scaffolding**: "스캐폴딩" 으로 번역합니다.
+* **trait**: "트레이트-trait"로 번역합니다.
+* **type-hint**: "타입 힌트" 라고 번역합니다.
+    * 동사형태로 사용된 경우 "type-hinting" "타입힌팅" 으로 번역합니다.
+* **token**: "토큰" 으로 번역합니다.
+* **validation**: 유효성 검증이라고 번역합니다.
+    * 경우에 따라 "입력값 검증" 으로 번역할 수도 있습니다.
+
+**고유명사들 (라이브러리, 패키지 등 네이밍된 기능들 - 영문 그대로 사용합니다.)**
+
+* Artisan
+* Blade
+* Cashier
+* Dusk
+* Eloquent
+* Envoy
+* Fortify
+* Homestead
+* Horizon
+* Livewire
+* Mix
+* Octane
+* Pint
+* Sail
+* Sanctum
+* Scout
+* Socialite
+* Stripe (외부 라이브러리이지만 대명사 이므로 단독 영문 표기)
+* Telescope
+* Valet
+* Vite
+
+# 안티패턴
+
+* "종속성 주입" → "**의존성 주입**"
+* “laravel” → “**라라벨**”
+    * Laravel은 → **라라벨은**
+    * Laravel에 → **라라벨에**
+    * Laravel의 → **라라벨의**
+    * Laravel에서 → **라라벨에서**
+    * Laravel에는 → **라라벨에는**
+    * Laravel → **라라벨**
+    * \[Laravel → **\[라라벨**
+* “메소드” → “**메서드**”
+* “응용 프로그램” → "**애플리케이션**"
+* “어플리케이션” → "**애플리케이션**”
+* “리디렉션" → “**리다이렉션**"
+* “도우미 함수" → "**헬퍼 함수**"
+* “암시적" → "**암묵적**"
+* warning 경고 → “**주의**"
+* "구성하기" → "**설정하기**"
+* “대기열" → “**queue-큐**”
